Create guests with kvm-install-vm
A script to create guests on a local kvm hypervisor from cloud-init
images. This is wrapper for the kvm-install-vm shell script.

A simple config file can be used to define sets of guests to be created
and destroyed.  The last mac address assigned is saved for the next
generation.

Also import an ansible playbook to setup the kvm hypervisor on a new
machine, using ansible and some roles from ansible-galaxy.