Add post create and distro command options
Instead of hard coding the systemd-resolve command after creating the
guests, move the command to run to the config.  Support expanding the
command string so we do not have to duplicate values, for example the
domain and bridge name for the systemd-resolve command.

For balance, add a post destroy command too.