Add example cloud-init userdata scripts directory
The scriptname option can be specified to have cloud-init run custom
setup scripts on first boot. Start a directory of examples that may be
useful.

Add the first example script called `apt-install-python.sh` which will
install python2 on debian/ubuntu. This can be helpful for older versions
of ansible (unless one messes with anisble_python_interpreter variables
for just the hosts that need it.)

    [ubuntu]
    desc = ubuntu test environment
    guests = 6
    distro = ubuntu1804
    scriptname = ~/virt/scripts/apt-install-python.sh