Fix autostart option
Set guests to autostart by default. This can be overridden in the config
file for a lab (or in the global section) with the autostart = no
option.  Remove the virt-lab create --autostart flag since this defaults
to true. The presence of that flag in the syntax does not us set the
autostart to be true by default (we could add a --no-autostart if we
wanted to override with a command line option).