Add support to run an ansible playbook on the created guests
Add support to run an ansible playbook on the guests created by virt-lab
create to do any customization of the newly created guests.  The
playbook will run within the postcreate process.  Run virt-lab create
again to rerun the playbook.

Add the --format inventory option to virt-lab list to provide the
dynamic inventory for ansible.  Add the script/inventory.sh wrapper for
ansible, to appease the particular command line syntax ansible wants for
inventory scripts.

Setup:
- Copy the postcreate.sh and inventory.sh scripts to ~/virt/scripts
- Copy the welcome.yaml to ~/virt/playbooks
- Setup the postcreate option(s) to run the postcreate script(s)
  in your virt-lab.cfg file. See the example cfg file.

A different path to the scripts can be setup by defining the scriptdir
in the virt-lib.cfg file.