Support ansible host groups in dynamic inventory
Support ansible host groups by interpreting a special sub-section in the
config file called <name>.group. This special section, if present, can
be used to organize the hosts into groups in the dynamic inventory
script output.

We use host groups in the OpenAFS roles (ansible-openafs) to set which
hosts are clients, fileserver, dbservers, and the kdc. So, with this
change it is now possible to use 'virt-lab create <name>' to create a
set of guests and automatically run a playbook to deploy a full cell.