Conditionaly run systemd-resolve based on command line options
Change the defaults of the 'gateway' and 'bridge' options to None to
allow one to omit them from the config file but set them on with command
line options.

Conditionaly run systemd-resolve in the local-dns.yaml playbook so we
can just aways run this playbook and skip the command when the gateway
and/or brigde is not defined in the config or the command line.