Read site-local options from ~/.virt-lab.cfg
Read site-local values from the ~/.virt-lab.cfg file. This is intended
as a convenience for setting values specific to a particular host, such
as the key, bridge, and gateway values when we want to share an
otherwise host independent set of labs, instead of needed to pass those
as command line arguments.