Expand all strings, except namefmt
Expand all of the strings and not just the postcreate and postdestroy
options, with the exception of 'namefmt' which is meant to be a format
string.  This allows us to define options which can be used as
variables.

Also set all None values to empty strings so we can effectively use
the empty values as placeholders.

Be sure to set kvm-install-vm args when a value is present.